firebase workinggit add --all!
