+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^4.6.0",
     "firebase": "^8.4.3",
+    "firebase-admin": "^9.7.0",
     "react": "^17.0.2",
     "react-bootstrap": "^1.5.2",
     "react-dom": "^17.0.2",
@@ -18,9 +19,11 @@
     "web-vitals": "^1.1.1"
   },
   "scripts": {
+    "predeploy": "npm run build",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "fbdeploy": "npm run predeploy && firebase deploy",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
